+++ src/actions/index.js
@@ -10,13 +10,13 @@ export {
   copy,
   copyContent,
   cut,
-  moveContent
+  moveContent,
 } from '@plone/volto/actions/clipboard/clipboard';
 export {
   addComment,
   deleteComment,
   listComments,
-  updateComment
+  updateComment,
 } from '@plone/volto/actions/comments/comments';
 export {
   createContent,
@@ -25,11 +25,11 @@ export {
   orderContent,
   resetContent,
   sortContent,
-  updateContent
+  updateContent,
 } from '@plone/volto/actions/content/content';
 export {
   getContentTypeTypes,
-  updateContentTypeFieldTypes
+  updateContentTypeFieldTypes,
 } from '@plone/volto/actions/contenttype/contenttype';
 export {
   deleteControlpanel,
@@ -38,42 +38,47 @@ export {
   getSystemInformation,
   listControlpanels,
   postControlpanel,
-  updateControlpanel
+  updateControlpanel,
 } from '@plone/volto/actions/controlpanels/controlpanels';
 export { getDiff } from '@plone/volto/actions/diff/diff';
 export { emailNotification } from '@plone/volto/actions/emailNotification/emailNotification';
 export {
   getFieldSchema,
-  updateFieldSchema
+  updateFieldSchema,
 } from '@plone/volto/actions/fieldschema/fieldschema';
 export {
   createGroup,
   deleteGroup,
   getGroup,
   listGroups,
-  updateGroup
+  updateGroup,
 } from '@plone/volto/actions/groups/groups';
 export {
   getHistory,
-  revertHistory
+  revertHistory,
 } from '@plone/volto/actions/history/history';
 export {
   addMessage,
   purgeMessages,
-  removeMessage
+  removeMessage,
 } from '@plone/volto/actions/messages/messages';
 export { getNavigation } from '@plone/volto/actions/navigation/navigation';
 export { getQuerystring } from '@plone/volto/actions/querystring/querystring';
 export { getQueryStringResults } from '@plone/volto/actions/querystringsearch/querystringsearch';
 export { listRoles } from '@plone/volto/actions/roles/roles';
-export { getSchema } from '@plone/volto/actions/schema/schema';
+export {
+  getSchema,
+  postSchema,
+  putSchema,
+  updateSchema,
+} from '@plone/volto/actions/schema/schema';
 export {
   resetSearchContent,
-  searchContent
+  searchContent,
 } from '@plone/volto/actions/search/search';
 export {
   getSharing,
-  updateSharing
+  updateSharing,
 } from '@plone/volto/actions/sharing/sharing';
 export { setSidebarTab } from '@plone/volto/actions/sidebar/sidebar';
 export { setExpandedToolbar } from '@plone/volto/actions/toolbar/toolbar';
@@ -87,25 +92,24 @@ export {
   resetPassword,
   setInitialPassword,
   updatePassword,
-  updateUser
+  updateUser,
 } from '@plone/volto/actions/users/users';
 export {
   login,
   loginRenew,
-  logout
+  logout,
 } from '@plone/volto/actions/userSession/userSession';
 export {
   getVocabulary,
-  getVocabularyTokenTitle
+  getVocabularyTokenTitle,
 } from '@plone/volto/actions/vocabularies/vocabularies';
 export {
   getWorkflow,
-  transitionWorkflow
+  transitionWorkflow,
 } from '@plone/volto/actions/workflow/workflow';
 export {
   installAddon,
   listAddons,
   uninstallAddon,
-  upgradeAddon
+  upgradeAddon,
 } from './addons/addons';
-