+++ src/components/manage/Blocks/Image/Edit.jsx
@@ -219,7 +219,9 @@ class Edit extends Component {
    */
   render() {
     const { data } = this.props;
-
+    const placeholder =
+      this.props.data.placeholder ||
+      this.props.intl.formatMessage(messages.ImageBlockInputPlaceholder);
     return (
       <div
         className={cx(
@@ -293,9 +295,7 @@ class Edit extends Component {
                     <Input
                       onKeyDown={this.onKeyDownVariantMenuForm}
                       onChange={this.onChangeUrl}
-                      placeholder={this.props.intl.formatMessage(
-                        messages.ImageBlockInputPlaceholder,
-                      )}
+                      placeholder={placeholder}
                       value={this.state.url}
                       // Prevents propagation to the Dropzone and the opening
                       // of the upload browser dialog