+++ src/components/manage/Blocks/Title/Edit.jsx
@@ -136,6 +136,11 @@ class Edit extends Component {
     if (__SERVER__) {
       return <div />;
     }
+
+    const placeholder =
+      this.props.data.placeholder ||
+      this.props.intl.formatMessage(messages.title);
+
     return (
       <Editor
         onChange={this.onChange}
@@ -150,7 +155,7 @@ class Edit extends Component {
           );
           return 'handled';
         }}
-        placeholder={this.props.intl.formatMessage(messages.title)}
+        placeholder={placeholder}
         blockStyleFn={() => 'documentFirstHeading'}
         onUpArrow={() => {
           const selectionState = this.state.editorState.getSelection();