+++ src/components/manage/Controlpanels/ContentTypes.jsx
@@ -2,29 +2,35 @@
  * Content Types component.
  * @module components/manage/Controlpanels/ContentTypes
  */
-
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
+import { compose } from 'redux';
+import { Link } from 'react-router-dom';
+import { getId, getParentUrl } from '@plone/volto/helpers';
+import { Portal } from 'react-portal';
+import { last } from 'lodash';
+import { Confirm, Container, Table, Button, Dropdown } from 'semantic-ui-react';
+import { toast } from 'react-toastify';
+import { FormattedMessage, defineMessages, injectIntl } from 'react-intl';
+import {
+  Error,
+  Icon,
+  ModalForm,
+  Toolbar,
+  Toast,
+} from '@plone/volto/components';
 import {
   deleteControlpanel,
   getControlpanel,
   postControlpanel,
 } from '@plone/volto/actions';
-import { Icon, ModalForm, Toast, Toolbar } from '@plone/volto/components';
-import { getId, getParentUrl } from '@plone/volto/helpers';
 import addSVG from '@plone/volto/icons/add-document.svg';
 import backSVG from '@plone/volto/icons/back.svg';
 import trashSVG from '@plone/volto/icons/delete.svg';
 import folderSVG from '@plone/volto/icons/folder.svg';
 import editSVG from '@plone/volto/icons/pen.svg';
-import { last } from 'lodash';
-import PropTypes from 'prop-types';
-import React, { Component } from 'react';
-import { defineMessages, FormattedMessage, injectIntl } from 'react-intl';
-import { Portal } from 'react-portal';
-import { connect } from 'react-redux';
-import { Link } from 'react-router-dom';
-import { toast } from 'react-toastify';
-import { compose } from 'redux';
-import { Button, Confirm, Container, Dropdown, Table } from 'semantic-ui-react';
+import layoutSVG from '@plone/volto/icons/file.svg';
 
 const messages = defineMessages({
   add: {
@@ -98,14 +104,7 @@ class ContentTypes extends Component {
     postControlpanel: PropTypes.func.isRequired,
     deleteControlpanel: PropTypes.func.isRequired,
     pathname: PropTypes.string.isRequired,
-    postRequest: PropTypes.shape({
-      loading: PropTypes.bool,
-      loaded: PropTypes.bool,
-    }).isRequired,
-    deleteRequest: PropTypes.shape({
-      loading: PropTypes.bool,
-      loaded: PropTypes.bool,
-    }).isRequired,
+    cpanelRequest: PropTypes.objectOf(PropTypes.any).isRequired,
     controlpanel: PropTypes.shape({
       '@id': PropTypes.string,
       items: PropTypes.arrayOf(
@@ -132,6 +131,7 @@ class ContentTypes extends Component {
     this.onAddTypeSuccess = this.onAddTypeSuccess.bind(this);
     this.onEdit = this.onEdit.bind(this);
     this.addFields = this.addFields.bind(this);
+    this.onLayout = this.onLayout.bind(this);
     this.onDelete = this.onDelete.bind(this);
     this.onDeleteCancel = this.onDeleteCancel.bind(this);
     this.onDeleteOk = this.onDeleteOk.bind(this);
@@ -142,6 +142,7 @@ class ContentTypes extends Component {
       addTypeError: '',
       showDelete: false,
       typeToDelete: undefined,
+      error: null,
       isClient: false,
     };
   }
@@ -165,17 +166,36 @@ class ContentTypes extends Component {
   }
 
   UNSAFE_componentWillReceiveProps(nextProps) {
+    // Get
+    if (
+      this.props.cpanelRequest.get.loading &&
+      nextProps.cpanelRequest.get.error
+    ) {
+      this.setState({
+        error: nextProps.cpanelRequest.get.error,
+      });
+    }
+
     // Create
-    if (this.props.postRequest.loading && nextProps.postRequest.loaded) {
+    if (
+      this.props.cpanelRequest.post.loading &&
+      nextProps.cpanelRequest.post.loaded
+    ) {
       this.props.getControlpanel(this.props.id);
       this.onAddTypeSuccess();
     }
-    if (this.props.postRequest.loading && nextProps.postRequest.error) {
-      this.onAddTypeError(nextProps.postRequest.error);
+    if (
+      this.props.cpanelRequest.post.loading &&
+      nextProps.cpanelRequest.post.error
+    ) {
+      this.onAddTypeError(nextProps.cpanelRequest.post.error);
     }
 
     // Delete
-    if (this.props.deleteRequest.loading && nextProps.deleteRequest.loaded) {
+    if (
+      this.props.cpanelRequest.delete.loading &&
+      nextProps.cpanelRequest.delete.loaded
+    ) {
       this.props.getControlpanel(this.props.id);
       this.onDeleteTypeSuccess();
     }
@@ -242,10 +262,18 @@ class ContentTypes extends Component {
     this.props.history.push(value);
   }
 
+  /**
+   * Layout button click
+   * @param {*} event
+   * @param {string} value
+   * @returns {undefined}
+   */
+  onLayout(event, { value }) {
+    this.props.history.push(value);
+  }
+
   /** Delete */
   /**
-   *
-   *
    * @param {*} event Event object.
    * @param {*} { value }
    * @memberof ContentTypes
@@ -322,6 +350,11 @@ class ContentTypes extends Component {
    * @returns {string} Markup for the component.
    */
   render() {
+    // Error
+    if (this.state.error) {
+      return <Error error={this.state.error} />;
+    }
+
     if (!this.props.controlpanel) {
       return <div />;
     }
@@ -356,7 +389,7 @@ class ContentTypes extends Component {
             submitError={this.state.addTypeError}
             onCancel={() => this.setState({ showAddType: false })}
             title={this.props.intl.formatMessage(messages.addTypeFormTitle)}
-            loading={this.props.postRequest.loading}
+            loading={this.props.cpanelRequest.post.loading}
             schema={{
               fieldsets: [
                 {
@@ -424,11 +457,15 @@ class ContentTypes extends Component {
                       <Table.Cell>{item.description}</Table.Cell>
                       <Table.Cell>{item.count}</Table.Cell>
                       <Table.Cell textAlign="right">
-                        <Dropdown icon="ellipsis horizontal">
+                        <Dropdown
+                          icon="ellipsis horizontal"
+                          className={`actions-${item.id}`}
+                        >
                           <Dropdown.Menu className="left">
                             <Dropdown.Item
                               onClick={this.onEdit}
                               value={`${this.props.pathname}/${item['id']}`}
+                              className={`edit-${item.id}`}
                             >
                               <Icon name={editSVG} size="15px" />
                               <FormattedMessage
@@ -437,8 +474,20 @@ class ContentTypes extends Component {
                               />
                             </Dropdown.Item>
                             <Dropdown.Item
+                              onClick={this.onLayout}
+                              value={`${this.props.pathname}/${item.id}/layout`}
+                              className={`layout-${item.id}`}
+                            >
+                              <Icon name={layoutSVG} size="15px" />
+                              <FormattedMessage
+                                id="Layout"
+                                defaultMessage="Layout"
+                              />
+                            </Dropdown.Item>
+                            <Dropdown.Item
                               onClick={this.onDelete}
                               value={item['@id']}
+                              className={`delete-${item.id}`}
                             >
                               <Icon name={trashSVG} size="15px" />
                               <FormattedMessage
@@ -512,8 +561,7 @@ export default compose(
   connect(
     (state, props) => ({
       controlpanel: state.controlpanels.controlpanel,
-      postRequest: state.controlpanels.post,
-      deleteRequest: state.controlpanels.delete,
+      cpanelRequest: state.controlpanels,
       pathname: props.location.pathname,
       id: last(props.location.pathname.split('/')),
     }),