+++ src/components/manage/Form/Form.jsx
@@ -110,6 +110,7 @@ class Form extends Component {
     submitLabel: PropTypes.string,
     resetAfterSubmit: PropTypes.bool,
     isEditForm: PropTypes.bool,
+    isAdminForm: PropTypes.bool,
     title: PropTypes.string,
     error: PropTypes.shape({
       message: PropTypes.string,
@@ -134,6 +135,7 @@ class Form extends Component {
     submitLabel: null,
     resetAfterSubmit: false,
     isEditForm: false,
+    isAdminForm: false,
     title: null,
     description: null,
     error: null,
@@ -247,7 +249,7 @@ class Form extends Component {
   }
 
   hideHandler = (data) => {
-    return !blockHasValue(data);
+    return !!data.fixed || !blockHasValue(data);
   };
 
   /**
@@ -820,6 +822,7 @@ class Form extends Component {
                               pathname={this.props.pathname}
                               block={block}
                               selected={this.state.selected === block}
+                              manage={this.props.isAdminForm}
                             />
                           </div>
                         </div>